ui: move column code into its own file.
